Added test to show that upstream synch works
