+++ chime/worker.py
@@ -10,11 +10,9 @@ import traceback
 
 from git import Repo
 
-from . import NEEDS_PUSH_FILE
-from .view_functions import _remote_exists
+from .repo_functions import NEEDS_PUSH_FILE, push_upstream_if_needed
 from .google_api_functions import (
-    is_overdue_ga_config, read_ga_config, request_new_google_access_token,
-    WriteLocked
+    is_overdue_ga_config, read_ga_config, request_new_google_access_token
 )
 
 parser = argparse.ArgumentParser(description='Do the things that need doing.')
@@ -44,15 +42,7 @@ if __name__ == '__main__':
             traceback.print_exc(file=sys.stderr)
         
         try:
-            needs_push_file = os.path.join(running_state_dir, NEEDS_PUSH_FILE)
-
-            if os.path.exists(needs_push_file):
-                with WriteLocked(needs_push_file) as file:
-                    repo = Repo(repo_path)
-                    if _remote_exists(repo, 'origin'):
-                        Logger.info('  pushing origin {}'.format(repo))
-                        repo.git.push('origin', all=True, with_exceptions=True)
-                    os.remove(file.name)
+            push_upstream_if_needed(Repo(repo_path), running_state_dir)
         except:
             traceback.print_exc(file=sys.stderr)
